requirements.txt
Django>=3.2
django-debug-toolbar>=3.2.1
djangorestframework>=3.12.4